documentacao/requirements.txt
flask
pyodbc
flask_sqlalchemy
python-dotenv
plotly
pandas
flask-login
flask-bcrypt
wfastcgi
os-sys